:dart: Improve: correcciones en post, agregar tipado y documentacion

+++ src/modules/post/post.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 
 import Post from 'src/entity/post.entity';
 import { CrearPostDTO } from './dtos/create-post.dto';
+import { UpdatePostDTO } from './dtos/update-post.dto';
 
 @Injectable()
 export class PostService {
@@ -24,6 +25,14 @@ export class PostService {
         pst_imagen_4,
         pst_imagen_5,
       } = newpost;
+
+      const check = await this.postRepository.findOne({ pst_descripcion });
+      if (check) {
+        throw new HttpException(
+          { status: HttpStatus.BAD_REQUEST, error: 'Post ya existe' },
+          HttpStatus.BAD_REQUEST,
+        );
+      }
 
       const post = this.postRepository.create({
         pst_isActive: true, // La activacion va a estar por defecto
@@ -57,43 +66,71 @@ export class PostService {
 
   async getPost(id: number) {
     try {
-      return await this.postRepository.findOne(id);
+      const post = await this.postRepository.findOne(id);
+
+      if (!post) {
+        throw new HttpException(
+          { status: HttpStatus.BAD_REQUEST, error: 'Usuario ya existe' },
+          HttpStatus.BAD_REQUEST,
+        );
+      }
+      return post;
     } catch (error) {
       return error;
     }
   }
 
-  async updatePost(id: number, newpost: CrearPostDTO) {
+  async updatePost(id: number, newpost: UpdatePostDTO) {
     try {
+      const {
+        pst_isActive,
+        pst_descripcion,
+        pst_imagen_1,
+        pst_imagen_2,
+        pst_imagen_3,
+        pst_imagen_4,
+        pst_imagen_5,
+        pst_precioBase,
+      } = newpost;
+
       const post = await this.postRepository.findOne(id);
 
-      // No cambio el id de usuario, porq se supone q es el mismo usuario (creo)
-      post.pst_descripcion = newpost.pst_descripcion;
-      post.pst_imagen_1 = newpost.pst_imagen_1;
-      post.pst_imagen_2 = newpost.pst_imagen_2;
-      post.pst_imagen_3 = newpost.pst_imagen_3;
-      post.pst_imagen_4 = newpost.pst_imagen_4;
-      post.pst_imagen_5 = newpost.pst_imagen_5;
-      post.pst_precioBase = newpost.pst_precioBase;
+      if (!post) {
+        throw new HttpException(
+          { status: HttpStatus.BAD_REQUEST, error: 'Usuario ya existe' },
+          HttpStatus.BAD_REQUEST,
+        );
+      }
 
-      return await this.postRepository.save(post);
-    } catch (error) {
-      return error;
-    }
-  }
+      if (pst_descripcion) {
+        post.pst_descripcion = pst_descripcion;
+      }
+      if (pst_imagen_1) {
+        post.pst_imagen_1 = pst_imagen_1;
+      }
+      if (pst_imagen_2) {
+        post.pst_imagen_2 = pst_imagen_2;
+      }
+      if (pst_imagen_3) {
+        post.pst_imagen_3 = pst_imagen_3;
+      }
+      if (pst_imagen_4) {
+        post.pst_imagen_4 = pst_imagen_4;
+      }
+      if (pst_imagen_5) {
+        post.pst_imagen_5 = pst_imagen_5;
+      }
+      if (pst_precioBase) {
+        post.pst_precioBase = pst_precioBase;
+      }
+      if (pst_isActive) {
+        post.pst_isActive = pst_isActive;
+      }
 
-  // Cambia el estado a habilitado o deshabilitado
-  async changeStatus(id: number, status: boolean) {
-    try {
-      const post = await this.postRepository.findOne(id);
-
-      post.pst_isActive = status;
-
-      await this.postRepository.save(post);
+      await post.save();
 
       return {
-        message: 'Se cambio el estado del post',
-        status,
+        message: 'El post se actualizo correctamente',
       };
     } catch (error) {
       return error;
@@ -102,9 +139,19 @@ export class PostService {
 
   async deletePost(id: number) {
     try {
+      const check = await this.postRepository.findOne(id);
+
+      if (!check) {
+        throw new HttpException(
+          { status: HttpStatus.BAD_REQUEST, error: 'Usuario ya existe' },
+          HttpStatus.BAD_REQUEST,
+        );
+      }
+
       await this.postRepository.delete(id);
+
       return {
-        message: 'Se elimino el post',
+        message: 'Se elimino el post correctamente',
       };
     } catch (error) {
       return error;
